% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOY_MatingPairSumm.R
\name{AMOY_MatingPairSumm}
\alias{AMOY_MatingPairSumm}
\title{Return end of season AMOY mating pair summary}
\usage{
AMOY_MatingPairSumm(x, ODBC_connect = TRUE, export = FALSE)
}
\arguments{
\item{x}{Denote "x" in parentheses to return a \code{data.frame} of all AMOY observations.}
}
\value{
This function returns the end of season AMOY mating pair summary as a \code{data.frame}.
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB and returns the end of season AMOY mating pair summary.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
AMOY_MatingPairSumm(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
