% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOYPairsByDate.R
\name{AMOYPairsByDate}
\alias{AMOYPairsByDate}
\title{AMOYPairsByDate}
\usage{
AMOYPairsByDate(x)
}
\arguments{
\item{x}{Denote "x" in parentheses to return a \code{data.frame} of a summary of AMOY mating pairs.}
}
\value{
Returns counts of the number of mating AMOY pairs per event as a \code{list}; the first element for plotting and the second for tabular display.
}
\description{
This script summarizes the AMOY mating pair survey data from \code{\link{GetAMOYData}} by Date for plotting and analysis.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
 
AMOYPairsByDate(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
