% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEvents.R
\name{GetEvents}
\alias{GetEvents}
\title{Return in-season survey events for each survey type from the Coastal birds Access database.}
\usage{
GetEvents(connect = "ODBC", DBfile = NULL, export = FALSE)
}
\arguments{
\item{connect}{Should the function connect to the Access DB? The default 
(\code{connect = `ODBC`}) is to try to connect using the Windows ODBC manager. 
If the connection is not available or not desired, one can use \code{connect = `Hmisc`}
and include a patch to a saved version of the database, or
the function can return the saved data from the package (\code{connect = `No`}). 
Note the saved data may not be up-to-date.}

\item{DBfile}{Path to a specified database file.}

\item{export}{Should the incubation data be exported as a csv file and RData object?
(This argument is used to regenerate the RData for the package.)}
}
\value{
This function returns all of the event level data for each survey type as a \code{data.frame}.
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB 
(Access backend entered as 'NETNCB' in Windows ODBC manager) and returns summary information
on surveys
}
\examples{
# events <- GetEvents()
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
