% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIncubationData.R
\name{GetIncubationData}
\alias{GetIncubationData}
\title{Return incubation surveys from database}
\usage{
GetIncubationData(x, ODBC_connect = TRUE, export = FALSE)
}
\arguments{
\item{ODBC_connect}{Should the function connect to the Access DB? The default (TRUE) is to 
try to connect using the Windows ODBC manager. If the connection is not available or not desired, 
the function can return the saved data from the package. 
Note the saved data may not be up-to-date.}

\item{export}{Should the incubation data be exported as a csv file and RData object?
(This argument is used to regenerate the RData for the package.)}
}
\value{
This function returns the raw boat-based incubation survey data as a \code{data.frame}.
}
\description{
This function connects to the backend of NETN's Coastal Bird 
Access DB (Access backend entered as 'NETNCB' in Windows ODBC manager) and 
returns the boat-based incubation data. If the Access DB is not
accessible from the ODBC connection, the function returns a saved image of the data.
}
\examples{
incubation <- GetIncubationData(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
