% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyMat.R
\name{GetSurveyMat}
\alias{GetSurveyMat}
\title{View survey matrix}
\usage{
GetSurveyMat(
  survey = NA,
  df = NULL,
  island = NA,
  year = NA,
  species = NA,
  time = "year"
)
}
\arguments{
\item{survey}{A character vector. Accepts "Nest", "Creche", or "Incubation". There is no default and this must be denoted.}

\item{df}{A data frame of the surveys. Supply argument when connection to database 
is not available. If survey is "creche", then must be data frame created from 
\code{GetCrecheData} function. If survey is "Incubation", then must be data frame created from the
\code{GetIncubationData} function. If survey is "Nest", then must be data frame created from the
\code{GetNestData} function.}

\item{island}{A  vector of island names. To view summaries across all islands, 
"All Islands"}

\item{year}{Calendar year(s) to view effort by. Useful when wanting to view seasonal surveys with time = "Date".
survey data in a year.}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{time}{Select whether to show effort by year ("year) or for individual dates ("Date"). Defaults to "year".}
}
\value{
This function returns a matrix showing dates of surveys for 
specified arguments indicated as "X".
}
\description{
Constructs matrix showing history of survey for particular surveys
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager if df = NULL.
}

\examples{
GetSurveyMat(survey ="Nest", species ="COEI", year = 2009, time="Date")
GetSurveyMat(survey ="Creche",  year = 2007:2011)
GetSurveyMat(survey ="Incubation",  year = 2007, time="Date")


}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
