% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCBBData.R
\name{importCBBData}
\alias{importCBBData}
\title{importCBBData Import NETN Coastal Bird data that are formatted as .xlxs files.}
\usage{
importCBBData(path = NA, new_env = TRUE, zip_name = NA)
}
\arguments{
\item{path}{Quoted path of folder containing tables.}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in CBB_TABLES environment. If \code{FALSE}, stores views in global environment}

\item{zip_name}{Quoted string ending in .zip. If specified, function looks for the specified file name and
imports .csvs from the zip file. If not specified, function looks for and imports individual csvs. Note that
this takes slightly longer than loading individual .csvs, due to the unzipping process.}
}
\value{
NETN database views in specified environment
}
\description{
This function imports all CSVs exported from NETN's Coastal Bird Database.
}
\examples{
\dontrun{
# Import individual xlsxs into global environment
importCBBData(path = "C:/Coastal_Birds/exports/NETN")

# Import zipped csvs into global environment
path <- "C:/Coastal_Birds/exports/NETN"
importCBBData(path = path, zip_name = "CBB_Dataset_Export_20230119.zip")
}

}
