% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestsByDate.R
\name{NestsByDate}
\alias{NestsByDate}
\title{NestsByDate}
\usage{
NestsByDate(x)
}
\arguments{
\item{x}{Denote in parentheses to return a summary of all ground-based surveys by species, island, across all islands, and date.}
}
\value{
Returns a \code{list}  with the counts of nests and thier contents (chicks or eggs). The first  \code{list}  element summarizes nest surveys by Date for graphing. The second element for tabular summary.
}
\description{
Brings in the raw ground-based nest survey data from \code{\link{GetNestData}} and summarizes it by date for plotting and analysis.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
NestsByDate(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
