% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Plot_notgeneric.R
\name{plotCB}
\alias{plotCB}
\title{CBPlot}
\usage{
plotCB(data, species = NA, island = NA, year = NA, scale = "norm",
  facet = "Island", var = NA)
}
\arguments{
\item{data}{A \code{data.frame}  of coastal bird observations summarized for plotting. Typically from \code{\link{AMOYPairsByDate}}, \code{\link{AMOYPairsByYear}}, \code{\link{CrecheSum}}, \code{\link{NestByDate}}, or \code{\link{NestsByYear}}}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{island}{A vector of island names (e.g., "Calf"). To view summariaes across all islands, "All Islands"}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{scale}{Convert to log scale by entering "log"}

\item{facet}{Plot the data into separate facets by Island, Species, etc.}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}
}
\value{
Outputs a ggplot graph of species detections over time.
}
\description{
Plots bird detections over time.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
plotCB(data= nests.date, species= "LETE",  facet= "Island", island = "Lovells",var = "Nests", year= 2010)
plotCB(data= nests.date, species= "BCNH", scale="norm" ,facet= "Island", var = "Eggs")
CBPlot(nests.year, scale = "log", facet= "Island", var = NA)
CBPlot(nests.date, scale = "norm", facet= "Island", var = NA, island = NA, year= 2008, species= "COEI")
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
