% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumCOTE.R
\name{SumCOTE}
\alias{SumCOTE}
\title{sum COTE incubation surveys}
\usage{
SumCOTE(df = NULL, time, output = "graph", ByObserver = "no")
}
\arguments{
\item{df}{The user can optionally load the raw incubation data from an R object or connect to the 
Access database to obtain it. Defaults to NULL, which means the Access database will
be used to obtain it.}

\item{time}{Character string equal to "date" or "year". Value must be provided; there is
no default. Choose to return raw counts by "date" or max counts per "year".}

\item{output}{Character string equal to "graph" or "table". 
Defaults to long format (output= "graph") ready for ggplot and the \code{\link{PlotBirds}}
function. For wide format use "table".}

\item{ByObserver}{Character string equal to "yes" or "no".  If "yes" will output the 
survey data counted by each observer for each island 
segment on each date. Only sums across multiple observations by same observer at each segment. 
Defaults to "no".}
}
\value{
Returns a \code{data.frame} with the counts of Common terns (COTE) observed during 
boat-based surveys
}
\description{
Brings in the raw incubation survey data from \code{\link{GetIncubationData}} and
 summarizes the data for plotting and analysis for Common Tern. Currently only returns counts from Spinnaker Platform.
}
\examples{
 
SumCOTE(time= "year", output = "graph")
SumCOTE(time= "date", output = "table")
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
