% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumGulls_DCCO.R
\name{SumGulls_DCCO}
\alias{SumGulls_DCCO}
\title{sum Gull and DCCO surveys}
\usage{
SumGulls_DCCO(
  time = "year",
  species = c("DCCO", "GBBG", "HERG"),
  output = "graph",
  ByObserver = "no",
  islands = c("Calf", "Little Calf", "Green", "The Graves", "Middle Brewster",
    "Outer Brewster", "Shag Rocks", "Little Brewster")
)
}
\arguments{
\item{time}{Character string equal to "date" or "year". Value must be provided; there is
no default. Choose to sum counts by "date" or "year". Summing by date will sum counts across 
segments of each island for each date. Summing by year sums counts across all surveys conducted 
in that year. Note that some surveys were repeated in the same year.}

\item{species}{To subset data by species, use "DCCO","GBBG","HERG". Defaults
to providing output for all species.}

\item{output}{Character string equal to "graph" or "table". 
Defaults to long format (output= "graph") ready for ggplot and the \code{\link{PlotBirds}}
function. For wide format use "table".}

\item{ByObserver}{Character string equal to "yes" or "no".  If "yes" will output the 
survey data counted by each observer for each island 
segment on each date. Only sums across multiple observations by same observer at each segment. 
Defaults to "no".}

\item{islands}{Defaults to summarizing counts only within the Outer Islands (Calf, Little Calf, Green,
The Graves, Middle Brewster, Outer Brewster, Shag Rocks and Little Brewster).}

\item{df}{The user can optionally load the raw incubation data from an R object or connect to the 
Access database to obtain it. Defaults to NULL, which means the Access database will
be used to obtain it.}
}
\value{
Returns a \code{list} with the counts of Gulls, cormorant, and terns observed during 
boat-based surveys per island, species, and date/year. The first  \code{list}  element summarizes 
incubation surveys by Year for graphing. The second element for tabular summary.
}
\description{
Brings in the raw incubation and Nest survey data from \code{\link{GetIncubationData}} 
and \code{\link{GetNestData}}, respectively, and summarizes the data for plotting and analysis. 
Currently only sums counts from the primary survey when repeated surveys were conducted. 
If you specify an argument to "ByObserver" this will return sum counts of all duplicate surveys by observer.
}
\examples{
 
SumGulls_DCCO(time= "year", species = "DCCO", output = "graph")
SumGulls_DCCO(time= "date", species = "DCCO", output = "graph")
SumGulls_DCCO(time= "date", ByObserver = "yes")

DCCO<-SumGulls_DCCO(time= "year", species= "DCCO")
PlotBirds(DCCO, var= "Incubating adults")
PlotBirds(DCCO, var= "Nests")
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
