% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumIncubation.R
\name{SumIncubation}
\alias{SumIncubation}
\title{sum incubation surveys}
\usage{
SumIncubation(df = NULL, time, species = NA, output = "graph",
  ByObserver = "no")
}
\arguments{
\item{df}{The user can optionally load the raw incubation data from an R object or connect to the 
Access database to obtain it. Defaults to NULL, which means the Access database will
be used to obtain it.}

\item{time}{Character string equal to "date" or "year". Value must be provided; there is
no default. Choose to sum counts by "date" or "year". Summing by date will sum counts across 
segments of each island for each date. Summing by year sums counts across all surveys conducted 
in that year. Note that some surveys were repeated in the same year.}

\item{species}{To subset data by species, use  "DCCO","GBBG","HERG". Defaults
to providing output for all species.}

\item{output}{Character string equal to "graph" or "table". 
Defaults to long format (output= "graph") ready for ggplot and the \code{\link{PlotBirds}}
function. For wide format use "table".
#' @param ByObserver Character string equal to \code{"yes"} or \code{"no"}.
If "yes" and \code{time}= "date", the function will output the survey data counted by each observer for 
each island segment on each date. Sums counts across multiple observations by same 
observer at each segment. Defaults to "no".}
}
\value{
Returns a \code{data.frame} with the raw and effort-adjusted counts of Gulls and DCCO incubating nests observed 
during boat-based incubation surveys per island, life stage, and time.
}
\description{
Brings in the raw incubation survey data from \code{\link{GetIncubationData}} and
 summarizes the data for plotting and analysis of DCCO and Gulls (no terns). Currently sums counts from the primary 
 survey conducted by the lead biologist when repeated surveys were conducted.  If you specify  \code{ByObserver}=\code{TRUE} 
  the  counts of all duplicate surveys will be summed by observer.
}
\section{Warning}{

Unless df is specified, the user must have an Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
 
SumIncubation(time= "year", species = "DCCO", output = "graph")
SumIncubation(time= "date", species = "DCCO", output = "graph")
SumIncubation(time= "date", ByObserver = "yes")
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
