% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumNestSurveys.R
\name{SumNestSurveys}
\alias{SumNestSurveys}
\title{Sum nest surveys}
\usage{
SumNestSurveys(time, species = NA, output = "graph")
}
\arguments{
\item{time}{Choose to sum counts by "date" or "year". Summing by date will sum counts across segments of each island for each date. Summing by year sums counts across all surveys conducted in that year. Note that some surveys were repeated in the same year.}

\item{species}{To subset data by species, use "BCNH" ,"COEI", "GLIB", "GREG", "SNEG", "DCCO", "GBBG", "HERG", "COTE" or "LETE"}

\item{output}{Defaults to long format ready for ggplot. For wide format use "table"}
}
\value{
Returns a \code{list}  with the counts of nests and thier contents (chicks or eggs). The first  \code{list}  element summarizes nest surveys by Date for graphing. The second element for tabular summary.
}
\description{
Brings in the raw ground-based nest survey data from \code{\link{GetNestData}} and summarizes it by date for plotting and analysis.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
SumNestSurveys(time= "date", species = "BCNH")
SumNestSurveys(time= "year", species = "BCNH")
SumNestSurveys(time= "year", output = "table")
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
