% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOYPairsByDate.R
\name{AMOYPairsByDate}
\alias{AMOYPairsByDate}
\title{AMOYPairsByDate}
\usage{
AMOYPairsByDate(x)
}
\arguments{
\item{survey}{Dataframe contructed from Access BE. Arguments can be "nest", "creche", "incubation" or "AMOY". If AMOY only returns dates when AMOY mating paris were reported}

\item{island}{A  vector of island names. To view summariaes across all islands, "All Islands"}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}
}
\description{
This script summarizes the AMOY mating pair survey data from \code{GetAMOYData} by Date for plotting and analysis.
}
\details{
# Return counts of the number of mating AMOY pairs per event
}
