% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Plot.R
\name{CBPlot}
\alias{CBPlot}
\title{CBPlot}
\usage{
CBPlot(object, species = NA, island = NA, var = NA, scale = NA,
  year = NA, facet = NA, ...)
}
\arguments{
\item{object}{A \code{data.frame}  of coastal bird observations}

\item{species}{A  vector of species names}

\item{island}{A  vector of island names}

\item{scale}{Convert to log scale by entering "log"}

\item{facet}{Plot the data into separate facets by Island, Species, etc.}

\item{time}{A numeric vector. Indicates which years should be graphed.}
}
\description{
Plots bird detections over time.
}
\details{
This function produces a graph of species detections over time.
}
