% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrecheSum.R
\name{CrecheSum}
\alias{CrecheSum}
\title{Sum creche surveys}
\usage{
CrecheSum(time, output = "graph", ByObserver = "no")
}
\arguments{
\item{time}{Choose to sum counts by "date" or "year". Summing by date will sum counts across segments of each island for each date. Summing by year sums counts across all surveys conducted in that year. Note that some surveys were repeated in the same year.}

\item{output}{Defaults to long format ready for ggplot. For wide format use "table"}

\item{ByObserver}{If "yes" will output the survey data counted by each observer for each island segment on each date. Only sums across multiple observations by same observer at each segment. Defaults to "no".}
}
\value{
Returns a \code{list} with the counts of COEI life stages observed during boat-based creche surveys per island, species, and by life stage. The first two \code{list}  elements summarize creche surveys by Date for tabular and graphing display, respectively. The 3rd and 4th \code{list} elements summarize creche surveys by YEAR for tabular and graphing display, respectively.
}
\description{
Brings in the raw creche survey data from \code{\link{GetCrecheData}} and summarizes the data by year or date for plotting and analysis.Currently only sums counts from the primary survey (Carol's) when repeated surveys were conducted. If you specify an argument to "ByObserver" this will return sum counts of all duplicate surveys by observer.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
 
CrecheSum(time ="date")
CrecheSum(time ="date", stage= "Chicks")
CrecheSum(time= "date", ByObserver = "yes")
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
