% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrecheSum.R
\name{CrecheSum}
\alias{CrecheSum}
\title{Sum creche surveys}
\usage{
CrecheSum(x)
}
\arguments{
\item{x}{Denote "x" in parentheses to return a summary of all creche surveys by island, across all islands, and date.}
}
\value{
Returns a \code{list} with the counts of COEI life stages observed during boat-based creche surveys per island, species, and by life stage. The first two \code{list}  elements summarize creche surveys by Date for tabular and graphing display, respectively. The 3rd and 4th \code{list} elements summarize creche surveys by YEAR for tabular and graphing display, respectively.
}
\description{
Brings in the raw creche survey data from \code{\link{GetCrecheData}} and summarizes the data by year and date for plotting and analysis
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
 
CrecheSum(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
