% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrecheSum.R
\name{CrecheSum}
\alias{CrecheSum}
\title{Sum creche surveys}
\usage{
CrecheSum(time, df = NULL, output = "graph", ByObserver = "no",
  islands = "outer")
}
\arguments{
\item{time}{Choose to sum counts by "date" or "year". Summing by date will sum 
counts across segments of each island for each date. Summing by year sums counts 
across all surveys conducted in that year. Note that some sites were surveyed more than once in 
the same year (primarily terns COTE and LETE).}

\item{df}{The user can optionally load the raw creche data from an R object or connect to the 
Access database to obtain it. Defaults to NULL, which means the Access database will
be used to obtain it.}

\item{output}{Character string equal to "graph" or "table". 
Defaults to long format (output= "graph") ready for ggplot and the \code{\link{PlotBirds}}
function. For wide format use "table".}

\item{ByObserver}{Character string equal to \code{"yes"} or \code{"no"}.
If "yes" and \code{time} = "date", the function will output the survey data counted by each observer for 
each island segment on each date. Sums counts across multiple observations by same 
observer at each segment. Defaults to "no".}

\item{islands}{Defaults to summarizing counts only within the Outer Islands (Calf, Little Calf, Green,
The Graves, Middle Brewster, Outer Brewster, Shag Rocks and Little Brewster).}
}
\value{
Returns a \code{data.frame} with the raw and effort-adjusted counts of COEI life stages observed 
during boat-based creche surveys per island, life stage, and time.
}
\description{
Brings in the raw creche survey data from \code{\link{GetCrecheData}} 
and summarizes the counts by year or date and life stage for plotting and analysis. Currently sums counts from the primary 
 survey conducted by the lead biologist when repeated surveys were conducted.  If you specify  \code{ByObserver}=\code{TRUE} 
  the  counts of all duplicate surveys will be summed by observer.
}
\section{Warning}{

Unless df is specified, the user must have an Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
 
CrecheSum(time ="date")
CrecheSum(time ="year")
CrecheSum(time= "date", ByObserver = "yes")
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
