% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAMOYData.R, R/GetAMOYdata.R
\name{GetAMOYdata}
\alias{GetAMOYdata}
\alias{GetAMOYdata}
\title{GetAMOYdata}
\usage{
GetAMOYdata(x)

GetAMOYdata(x)
}
\arguments{
\item{x}{Denote in parentheses to return df}

\item{x}{Denote in parentheses to return df}
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB and returns the raw AMOY survey data

This function connects to the backend of NETN's Coastal Bird Access DB and returns the raw AMOY survey data
}
\details{
This function returns the raw AMOY survey data as a \code{data.frame}.

This function returns the raw AMOY survey data as a \code{data.frame}.
}
