% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAMOYData.R
\name{GetAMOYData}
\alias{GetAMOYData}
\title{GetAMOYData}
\usage{
GetAMOYData(x)
}
\arguments{
\item{x}{Denote in parentheses to return df}
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB and returns the raw AMOY survey data
}
\details{
This function returns the raw AMOY survey data as a \code{data.frame}.
}
