% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAMOYData.R
\name{GetAMOYData}
\alias{GetAMOYData}
\title{Return AMOY surveys from database}
\usage{
GetAMOYData(x)
}
\arguments{
\item{x}{Denote "x" in parentheses to return a \code{data.frame} of all AMOY observations.}
}
\value{
This function returns the raw AMOY survey data as a \code{data.frame}.
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB and returns the raw AMOY survey data.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
GetAMOYdata(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
