% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCrecheData.R
\name{GetCrecheData}
\alias{GetCrecheData}
\title{Return creche surveys from database}
\usage{
GetCrecheData(x)
}
\arguments{
\item{x}{Denote in parentheses to return df}
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB and returns the raw creche survey data of COEI
}
\details{
This function returns the raw AMOY survey data as a \code{data.frame}.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
GetCrecheData(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
