% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIncubationData.R
\name{GetIncubationData}
\alias{GetIncubationData}
\title{GetIncubationData}
\usage{
GetIncubationData(x)
}
\arguments{
\item{x}{Denote in parentheses to return all boat-based incubation surveys.}
}
\value{
This function returns the raw boat-based incubation survey data as a \code{data.frame}.
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB and returns the boat-based incubation data.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
GetIncubationData(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
