% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNestData.R
\name{GetNestData}
\alias{GetNestData}
\title{Return ground-based nest surveys from database}
\usage{
GetNestData(x)
}
\arguments{
\item{x}{Denote in parentheses to return all ground-based nest surveys.}
}
\value{
This function returns the raw nest survey data as a \code{data.frame}.
}
\description{
This function connects to the backend of NETN's Coastal Bird Access DB and returns the raw ground-based nest survey data for non-AMOY (BCNH COEI COTE DCCO GBBG GLIB GREG HERG LETE SNEG SPSA WILL).
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
GetNestData(x)
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
