% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyMat.R
\name{getSurveyMat}
\alias{getSurveyMat}
\title{getSurveyMat}
\usage{
getSurveyMat(survey, island = NA, year = NA, species = NA)
}
\arguments{
\item{survey}{Dataframe contructed from Access BE. Arguments can be "nest", "creche", "incubation" or "AMOY". If AMOY only returns dates when AMOY mating paris were reported}

\item{island}{A  vector of island names. To view summariaes across all islands, "All Islands"}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{species}{A  vector of species name codes, e.g. "BCNH"}
}
\description{
Constructs survey matrix to show effort based on the inputs
}
\details{
This function produces a graph of species detections over time.
}
