% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyMat.R
\name{GetSurveyMat}
\alias{GetSurveyMat}
\title{View survey matrix}
\usage{
GetSurveyMat(survey, df = NULL, island = NA, year = NA,
  species = NA, time = "year")
}
\arguments{
\item{survey}{A character vector. Accepts "Nest", "Creche", or "Incubation". There is no default and this must be denoted.}

\item{df}{A data frame of the surveys. Supply argument when connection to database 
is not available. If survey is "creche", then must be data frame created from 
CrecheSum() function. If survey is "incubation", then must be data frame created from the
SumIncubation() function. If survey is "nest", then must be data frame created from the
SumNestSurveys() function.}

\item{island}{A  vector of island names. To view summaries across all islands, 
"All Islands"}

\item{year}{Calendar year(s) to view effort by. Useful when wanting to view seasonal surveys with time = "Date".
survey data in a year.}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{time}{Select whether to show effort by year ("year) or for individual dates ("Date"). Defaults to "year".}
}
\value{
This function returns a matrix showing dates of surveys for 
specified arguments indicated as "X".
}
\description{
Constructs survey matrix to show effort based on the inputs
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
GGetSurveyMat(survey ="Nests", species ="COEI", year = 2009, time="Date")
GetSurveyMat(survey ="Creche",  year = 2007:2011)
GetSurveyMat(survey ="Incubation",  year = 2007, time="Date")


}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
