% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBirds.R
\name{PlotBirds}
\alias{PlotBirds}
\title{Plot coastal bird survey data}
\usage{
PlotBirds(data, raw_count = FALSE, species = NA, island = NA,
  year = NA, scale = "norm", facet = "Island", var = NA,
  overlay_spp = FALSE, print = "yes", plot_title = "yes",
  legend = FALSE, Y_scale = "fixed")
}
\arguments{
\item{data}{A \code{data.frame}  of coastal bird observations summarized for plotting. Typically from \code{\link{SumIncubation}}, \code{\link{CrecheSum}}, or \code{\link{SumNestSurveys}}.}

\item{raw_count}{\code{TRUE} or \code{FALSE}. Plot effort-adjusted (default) or raw count data?}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{island}{A vector of island names (e.g., "Calf"). To view surveys summed across all islands, use "All Islands". WHen using "All Islands" you need to supply agrument to facet (such as "variable").}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{scale}{Convert to log scale by entering "log".}

\item{facet}{Plot the data into separate facets by Island, Species, or variable. Deafults to Island.}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}

\item{overlay_spp}{Defaults to \code{FALSE}. Enter \code{TRUE} if you would like to overlay time series for each species. This is needed to properly plot data when the \code{data.frame} to be plotted contains multiple species. In cases where the \code{data.frame} has multiple life stages (e.g., nests), provide argument to \code{var} to plot only one life stage at a time}

\item{print}{To not print plot enter "no".}

\item{plot_title}{Add a caption to the plot? Defaults to "yes". Enter "no" for no caption.}

\item{legend}{Add legend. Defaults to \code{FALSE}.}

\item{Y_scale}{Should y-scale of individual facets be on the same scale (default, "fixed") or not ("free_y")?}
}
\value{
Outputs a ggplot graph of species detections over time.
}
\description{
Plots bird detections over time.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
#Incubation surveys by year
dcco <- SumIncubation(time = "year", species = "DCCO")
PlotBirds(dcco)

# Incubation surveys by year of multiple species
incub<-SumIncubation(time = "year", species = c("DCCO", "HERG", "GBBG"))
PlotBirds(incub, overlay_spp= TRUE, raw_count= FALSE)

# Incubation surveys by date to view repeat effort
gbbg <- SumIncubation(time = "date", species = "GBBG")
PlotBirds(gbbg, year= "2012", legend =TRUE)

# Creche surveys by date; typically to view efforts in a single season
creche <- CrecheSum(time ="date")
# View survey counts in 2018
PlotBirds(creche, year = "2018")
# surveys summed across all islands
PlotBirds(creche, year = "2018", island= "All Islands", facet= "variable", legend = TRUE)

# Nest surveys
nests <- SumNestSurveys(time= "year", species = "BCNH")# annual counts of BCNH
PlotBirds(nests, var = "Nests")
PlotBirds(nests, island = "All Islands", facet= "variable", raw_counts= FALSE)
PlotBirds(nests, var= "EggsPerNest", legend =TRUE)

# Nest surveys of all species
nests<-SumNestSurveys(time= "year")
PlotBirds(nests, var = "Nests", overlay_spp= TRUE)

}
\seealso{
\url{https://www.nps.gov/im/netn/coastal-birds.htm}
}
