% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBirds.R
\name{PlotBirds}
\alias{PlotBirds}
\title{Plot coastal bird survey data}
\usage{
PlotBirds(data, species = NA, island = NA, year = NA,
  scale = "norm", facet = "Island", var = NA, overlay_spp = FALSE,
  print = "yes", plot_title = "yes", legend = FALSE)
}
\arguments{
\item{data}{A \code{data.frame}  of coastal bird observations summarized for plotting. Typically from \code{\link{SumIncubation}}, \code{\link{CrecheSum}}, or \code{\link{SumNestSurveys}}.}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{island}{A vector of island names (e.g., "Calf"). To view surveys summed across all islands, use "All Islands". WHen using "All Islands" you need to supply agrument to facet (such as "variable").}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{scale}{Convert to log scale by entering "log".}

\item{facet}{Plot the data into separate facets by Island, Species, or variable. Deafults to Island.}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}

\item{overlay_spp}{Defaults to \code{FALSE}. Enter \code{TRUE} if you would like to overlay time series for each species. This is needed to properly plot data when the \code{data.frame} to be plotted contains multiple species. In cases where the \code{data.frame} has multiple life stages (e.g., nests), provide argument to \code{var} to plot only one life stage at a time}

\item{print}{To not print plot enter "no".}

\item{plot_title}{Add a caption to the plot? Defaults to "yes". Enter "no" for no caption.}

\item{legend}{Add legend. Defaults to \code{FALSE}.}
}
\value{
Outputs a ggplot graph of species detections over time.
}
\description{
Plots bird detections over time.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
