% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Plot_notgeneric.R, R/PlotBirds.R
\name{PlotBirds}
\alias{PlotBirds}
\alias{PlotBirds}
\title{PlotBirds}
\usage{
PlotBirds(data, species = NA, island = NA, year = NA, scale = "norm",
  facet = "Island", var = NA)

PlotBirds(data, species = NA, island = NA, year = NA, scale = "norm",
  facet = "Island", var = NA)
}
\arguments{
\item{data}{A \code{data.frame}  of coastal bird observations summarized for plotting. Typically from \code{\link{SumIncubationByDate}}, \code{\link{SumIncubationByYear}}, \code{\link{CrecheSum}}, \code{\link{NestByDate}}, or \code{\link{NestsByYear}}}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{island}{A vector of island names (e.g., "Calf"). To view summariaes across all islands, "All Islands"}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{scale}{Convert to log scale by entering "log"}

\item{facet}{Plot the data into separate facets by Island, Species, etc.}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}

\item{data}{A \code{data.frame}  of coastal bird observations summarized for plotting. Typically from \code{\link{SumIncubation}}, \code{\link{CrecheSum}}, or \code{\link{SumNestSurveys}}.}

\item{island}{A vector of island names (e.g., "Calf"). To view summariaes across all islands, "All Islands"}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}

\item{scale}{Convert to log scale by entering "log"}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{facet}{Plot the data into separate facets by Island, Species, etc.}
}
\value{
Outputs a ggplot graph of species detections over time.

Outputs a ggplot graph of species detections over time.
}
\description{
Plots bird detections over time.

Plots bird detections over time.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.


User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
plotCB(data= nests.date, species= "LETE",  facet= "Island", island = "Lovells",var = "Nests", year= 2010)
plotCB(data= nests.date, species= "BCNH", scale="norm" ,facet= "Island", var = "Eggs")
CBPlot(nests.year, scale = "log", facet= "Island", var = NA)
CBPlot(nests.date, scale = "norm", facet= "Island", var = NA, island = NA, year= 2008, species= "COEI")
nests.date<-Get
PlotBirds(data= nests.date, species= "LETE",  facet= "Island", island = "Lovells",var = "Nests", year= 2010)
PlotBirds(data= nests.date, species= "BCNH", scale="norm" ,facet= "Island", var = "Eggs")
PlotBirds(nests.year, scale = "log", facet= "Island", var = NA)
PlotBirds(nests.date, scale = "norm", facet= "Island", var = NA, island = NA, year= 2008, species= "COEI")

# Incubation surveys by year
herg<-SumIncubation(time = "year", species = "HERG")
PlotBirds(herg)
# Incubation surveys by date to view seasonal surveya
herg<-SumIncubation(time = "date", species = "HERG")
PlotBirds(herg, year= "2007")

}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}

\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
