% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBirds.R
\name{PlotBirds}
\alias{PlotBirds}
\title{Plot coastal bird survey data}
\usage{
PlotBirds(data, species = NA, island = NA, year = NA,
  scale = "norm", facet = "Island", var = NA, overlay_spp = FALSE,
  print = "yes", caption = "yes")
}
\arguments{
\item{data}{A \code{data.frame}  of coastal bird observations summarized for plotting. Typically from \code{\link{SumIncubation}}, \code{\link{CrecheSum}}, or \code{\link{SumNestSurveys}}.}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{island}{A vector of island names (e.g., "Calf"). To view surveys summed across all islands, use "All Islands". WHen using "All Islands" you need to supply agrument to facet (such as "variable").}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{scale}{Convert to log scale by entering "log".}

\item{facet}{Plot the data into separate facets by Island, Species, or variable. Deafults to Island.}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}

\item{overlay_spp}{Defaults to \code{FALSE}. Enter \code{TRUE} if you would like to overlay time series for each species. This is needed to properly plot data when the \code{data.frame} to be plotted contains multiple species. In cases where the \code{data.frame} has multiple life stages (e.g., nests), provide argument to \code{var} to plot only one life stage at a time}

\item{print}{To not print plot enter "no".}

\item{caption}{Add a caption to the plot? Defaults to \code{TRUE}}
}
\value{
Outputs a ggplot graph of species detections over time.
}
\description{
Plots bird detections over time.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
# Incubation surveys by year
dcco <- SumIncubation(time = "year", species = "DCCO")
PlotBirds(dcco)

# Incubation surveys by year of multiple species
incub<-SumIncubation(time = "year", species = c("DCCO", "HERG", "GBBG"))
PlotBirds(incub, overlay_spp= TRUE)

# Incubation surveys by date to view repeat effort
lete <- SumIncubation(time = "date", species = "COTE")
PlotBirds(lete, year= "2012")

# Creche surveys by date; typically to view efforts in a single season
creche <- CrecheSum(time ="date")
# View survey counts in 2018
PlotBirds(creche, year = "2018")
# surveys summed across all islands
PlotBirds(creche, year = "2018", island= "All Islands", facet= "variable")

# Nest surveys
nests <- SumNestSurveys(time= "year", species = "BCNH")# annual counts of BCNH
PlotBirds(nests, var = "Nests")
PlotBirds(nests, island = "All Islands", facet= "variable")

# Nest surveys of all species
nests<-SumNestSurveys(time= "year")
PlotBirds(nests, var = "Nests", overlay_spp= TRUE)

}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
