% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBirds.R
\name{PlotBirds}
\alias{PlotBirds}
\title{Plot coastal bird survey data}
\usage{
PlotBirds(data, raw_count = FALSE, species = NA, island = NA,
  year = NA, stat = NA, scale = "norm", facet = "Island",
  var = NA, overlay = "var", print = "yes", plot_title = "yes",
  legend = FALSE, Y_scale = "fixed")
}
\arguments{
\item{data}{A \code{data.frame}  of coastal bird observations summarized for plotting. Typically from \code{\link{SumIncubation}}, \code{\link{CrecheSum}}, or \code{\link{SumNestSurveys}}.}

\item{raw_count}{\code{TRUE} or \code{FALSE}. Plot effort-adjusted (default) or raw count data?}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{island}{A vector of island names (e.g., "Calf"). To view surveys summed across all islands, use 
"All Islands". When using "All Islands" you need to supply agrument to facet (such as "variable").}

\item{year}{Numeric. Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{stat}{Character. Specify type of statistic to plot. Defaults to plotting the sum, but you can plot other
annual statistics by setting to "mean","max", or "min" when the \code{df} is generated by \code{\link{SumIncubation}} or \code{\link{CrecheSum}} functions.}

\item{scale}{Convert to log scale by entering "log".}

\item{facet}{Plot the data into separate facets by Island, Species, or variable. Defaults to Island. If \code{overlay = "island"} set \code{facet} to "CommonName", "Species_Code" , "FullLatinName" or "variable".}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}

\item{overlay}{Character. Defaults to \code{"var"} which overlays data by lifestage. Enter \code{"species"} to overlay time series 
by species when \code{data} contains multiple species. Enter \code{"island"} to overlay data by island. When overlaying by island and 
when multiple species are in \code{data}, denote \code{facet} as "CommonName", "Species_Code" or "FullLatinName".
In cases where the \code{data} has multiple life stages (e.g., nests), provide argument to \code{var} to plot only one life stage at a time.}

\item{print}{To not print plot enter "no".}

\item{plot_title}{Add a caption to the plot? Defaults to "yes". Enter "no" for no caption.}

\item{legend}{Add legend. Defaults to \code{FALSE}.
#' @param Y_scale Should y-scale of individual facets be on the same scale (default, "fixed") or not ("free_y")?}
}
\value{
Outputs a ggplot graph of species detections over time.
}
\description{
Plots bird detections over time.
}
\section{Warning}{

User must have Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
#Incubation surveys by year
dcco <- SumIncubation(time = "year", species = "DCCO",stat ="mean")
PlotBirds(dcco)

# Incubation surveys by year of multiple species
incub<-SumIncubation(time = "year", species = c("DCCO", "HERG", "GBBG"))
PlotBirds(incub, overlay= "species", raw_count= FALSE, stat ="mean", plot_title ="no")
PlotBirds(incub, overlay= "island", raw_count= FALSE, stat ="mean", plot_title ="no", facet="CommonName", legend= T) 
# Incubation surveys by date to view repeat effort
gbbg <- SumIncubation(time = "date", species = "GBBG")
PlotBirds(gbbg, year= "2012", legend =TRUE, stat ="sum")

# Creche surveys by date; typically to view efforts in a single season
creche <- CrecheSum(time ="date")
# View survey counts in 2018
PlotBirds(creche, year = "2018", stat= "sum")
# surveys summed across all islands
PlotBirds(creche, year = "2018", island= "All Islands", facet= "variable", legend = TRUE, stat= "sum")

# Nest surveys
nests <- SumNestSurveys(time= "year", species = "BCNH")# annual counts of BCNH
PlotBirds(nests, var = "Nests")
PlotBirds(nests, island = "All Islands", facet= "variable", raw_counts= FALSE)
PlotBirds(nests, var= "EggsPerNest", legend =TRUE)

# Nest surveys of all species
nests<-SumNestSurveys(time= "year")
PlotBirds(nests, var = "Nests", overlay= "species")

}
\seealso{
\url{https://www.nps.gov/im/netn/coastal-birds.htm}
}
