% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Plot_notgeneric.R
\name{plotCB}
\alias{plotCB}
\title{CBPlot}
\usage{
plotCB(data, species = NA, island = NA, year = NA, scale = "norm",
  facet = "Island", var = NA)
}
\arguments{
\item{data}{A \code{data.frame}  of coastal bird observations}

\item{species}{A  vector of species name codes, e.g. "BCNH"}

\item{island}{A  vector of island names. To view summariaes across all islands, "All Islands"}

\item{year}{Calendar year(s) to view data by. Useful when wanting to view seasonal survey data in a year.}

\item{scale}{Convert to log scale by entering "log"}

\item{facet}{Plot the data into separate facets by Island, Species, etc.}

\item{var}{Select a variable to plot, typically a life stage (e.g., Eggs, Nests, Creche size). Defaults to all values.}
}
\description{
Plots bird detections over time.
}
\details{
This function produces a graph of species detections over time.
}
