% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrecheSum.R, R/SumCreche.R
\name{SumCreche}
\alias{SumCreche}
\title{Sum creche surveys}
\usage{
SumCreche(
  df = NULL,
  time,
  survey_data = SurveyEffortBySpecies,
  segment = FALSE,
  output = "graph",
  ByObserver = "no",
  islands = "outer"
)

SumCreche(
  df = NULL,
  time,
  survey_data = SurveyEffortBySpecies,
  segment = FALSE,
  output = "graph",
  ByObserver = "no",
  islands = "outer"
)
}
\arguments{
\item{df}{Requires dataframe exported from NETN's data package imported via \code{\link{importCBBData}} from view 'qry_Dataset_4_Survey_Creche'. If \code{df} 
is \code{NULL}, the user must have an Access backend entered as 'NETNCB' in Windows ODBC manager in order to import from \code{\link{GetCrecheData}}.}

\item{time}{Character string equal to "date" or "year". Summing by date will sum 
counts (and associated densities) across segments of each island for each date. Summing by year sums counts 
across all surveys conducted in that year. Note that some sites were surveyed more than once in 
the same year (primarily terns COTE and LETE).}

\item{segment}{Logical. To summarize data at the survey (island-segment) scale (\code{TRUE}) or island-scale (\code{FALSE})
Defaults to \code{FALSE}.}

\item{output}{Character string equal to "graph" or "table". 
Defaults to exporting a data frame in long format (output= "graph") ready for ggplot and \code{\link{PlotBirds}}
For wide format enter "table" but note this currently just returns raw count values per life stage, not density.}

\item{ByObserver}{Character string equal to \code{"yes"} or \code{"no"}.
If "yes" and \code{time} = "date", the function will output the survey data counted by each observer for 
each island segment on each date. Sums counts across multiple observations by same 
observer at each segment. Defaults to "no".}

\item{islands}{Character. Defaults to summarizing counts only within the Outer Islands (Calf, Little Calf, Green,
The Graves, Middle Brewster, Outer Brewster, Shag Rocks and Little Brewster) or a vector of island names can be provided.}
}
\value{
Returns a \code{data.frame} with the raw and effort-adjusted counts of COEI life stages observed 
during boat-based creche surveys per island, life stage, and time.

Returns a \code{data.frame} with the raw and effort-adjusted counts of COEI life stages observed 
during boat-based creche surveys per island, life stage, and time.
}
\description{
Brings in the raw creche survey data from \code{\link{GetCrecheData}} 
and summarizes life stage counts by year or date to support analysis and reporting as defined in ####. Raw counts are also converted to
densities.This function currently ony sums counts from the primary survey conducted by the lead biologist when multiple observer surveys were conducted.
 When you specify \code{ByObserver = TRUE} and \code{time = "date"} the  raw counts and associated densities of all surveys are summed by observer.

Imports raw creche survey data and summarizes life stage counts by year or date. Raw counts are also converted to
densities.This function currently only sums counts from the primary surveyor conducted by the lead biologist when multiple observer surveys were conducted.
 When you specify \code{ByObserver = TRUE} and \code{time = "date"} the  raw counts and associated densities of all surveys are summed by observer.
}
\section{Warning}{

Unless \code{df} is specified, the user must have an Access backend entered as 'NETNCB' in Windows ODBC manager.


Unless \code{df} is specified, the user must have an Access backend entered as 'NETNCB' in Windows ODBC manager.
}

\examples{
 
# CrecheSum(time ="date")
# CrecheSum(time ="year")
# CrecheSum(time= "date", ByObserver = "yes")
\dontrun{ 
importCBBData(path, zip_name, new_env= TRUE) #creates CBB_TABLES object
# CrecheSum(df=CBB_TABLES$qry_Dataset_4_Survey_Creche, time ="date")
# CrecheSum(df=CBB_TABLES$qry_Dataset_4_Survey_Creche, time ="year")
# CrecheSum(df=CBB_TABLES$qry_Dataset_4_Survey_Creche, time= "date", ByObserver = "yes")
}
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}

\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
