% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumNestSurveys.R
\name{SumNestSurveys}
\alias{SumNestSurveys}
\title{Sum nest surveys}
\usage{
SumNestSurveys(time, species = NA, output = "graph", df, segment = FALSE)
}
\arguments{
\item{time}{Character string equal to "date" or "year".
Choose to sum effort-adjusted counts (no. / km2) by "date" or "year". Summing by date will sum counts 
across segments of each island for each date. Summing by year sums counts across all surveys 
conducted in that year per island; however, since there are multiple surveys per year for terns COTE and LETE, the maximum nest count among surveys is returned per year.}

\item{species}{To subset data by species, use "BCNH" ,"COEI", "GLIB", "GREG", 
"SNEG", "DCCO", "GBBG", "HERG", "COTE" or "LETE". Defaults
to providing output for all species.}

\item{output}{Character string equal to "graph" or "table". 
Defaults to long format (output= "graph") ready for ggplot and the \code{\link{PlotBirds}}
function. For wide format use "table".}

\item{df}{Dataframe. Requires dataframe exported from NETN's data package imported via \code{\link{importCBBData}} from view "qry_Dataset_3_Survey_Nest".}

\item{segment}{Would you like to summarize data at the island-segment scale (\code{TRUE}) or island-scale \code{FALSE})? Defaults to \code{FALSE}.}
}
\value{
Returns a \code{data.frame} with the counts of nests and chicks or eggs adjusted for survey-area (e.g. nests per km2) per Island and time
and estimates the Eggs and Chicks per Nest per island and time. Life stages are denoted under 'variable' eg., variable == "Nests","Chicks","Eggs", ChicksPerNest", "EggsPerNest","ClutchPerNest".
}
\description{
Brings in the raw ground-based nest survey data from \code{\link{importCBBData}}
and summarizes it by date or year and island for plotting and analysis.
}
\examples{
\dontrun{
importCBBData(path, zip_name, new_env= TRUE) #creates CBB_TABLES object
SumNestSurveys(df= CBB_TABLES$qry_Dataset_3_Survey_Nest, time= "date", species = "BCNH")
SumNestSurveys(df= CBB_TABLES$qry_Dataset_3_Survey_Nest, time= "year", species = "BCNH")
SumNestSurveys(df= CBB_TABLES$qry_Dataset_3_Survey_Nest, time= "year", output = "table")
}
}
\seealso{
\url{ https://www.nps.gov/im/netn/coastal-birds.htm}
}
